Fix the bug of /opt/skyeye installation
Now we only permit SkyEye installed at /opt/skyeye since we write a fix value
for directory of module loading. Now we can get the value of skyeye installation
directory and to load module acoording to the value. With this patch, we can install SkyEye in random directory of the filesystem.

Signed-off-by: Michael.Kang <blackfin.kang@gmail.com>