Modify the implementation of get_elf_endian function.
When using the remote gdb to debug an elf file, we should know the endian
about it.The macro which supplied by bfd will return little-endian infomation
whether the elf file is little-endian or big-endian.
Reference the implementation of the command of readelf to implement our
target----get endian from elf file.We can find the implementation of readelf
from the source code of binutils.