Add exception support for running linux kernel
Copy e600_ppc_exception here for ppc_dyncom use, for there is a little
difference. We need to keep track of both pc and phys_pc in dyncom.
In another hand, e600_dyncom_ppc_exception can implement with llvm IR
later on. So we have to separate it with interpretation.

dec_io_do_cycle in dyncom is also different.We just decrease the tbl
and dec register when running out of cpu_run and detect excception.
If running cycles more than TIMEOUT_THRESHOLD(defined in skyeye_dyncom.h)
in cpu_run, we force it return.

Signed-off-by: lizhijie <lizhijie86su@gmail.com>
Signed-off-by: Michael.Kang <blackfin.kang@gmail.com>