change emit_store_pc
To store both effective pc and physics pc.

emit_store_pc_end_page is used to store the next pc when running
the last instruction of a page(The instruction is not an jump
instruction).If it is a jump instruction at the end of a page,next
pc will be stored in the instruction.

emit_store pc_cond is to store pc when cond is true, which is used
in conditional branch translation.

Signed-off-by: lizhijie <lizhijie86su@gmail.com>
Signed-off-by: Michael.Kang <blackfin.kang@gmail.com>